# Conflicts:
#	src/font/sdf2_table.rs

+++ src/hal/web/mod.rs
@@ -18,31 +18,31 @@ extern "C" {
     // #[wasm_bindgen]
     fn fillBackGround(canvas: &HtmlCanvasElement, ctx: &CanvasRenderingContext2d, x: u32, y: u32);
     // #[wasm_bindgen]
-    fn setFont(ctx: &CanvasRenderingContext2d, weight: u32, fontSize: u32, font: u32, strokeWidth: u8);
+    fn setFont(ctx: &CanvasRenderingContext2d, weight: u32, fontSize: u32, font: f64, strokeWidth: u8);
     // #[wasm_bindgen]
     fn drawCharWithStroke(ctx: &CanvasRenderingContext2d, ch_code: u32, x: u32, y: u32);
-	fn getGlobalMetricsHeight(name: u32, size: f32) -> f32;
+	fn getGlobalMetricsHeight(name: f64, size: f32) -> f32;
     // #[wasm_bindgen]
     fn drawChar(ctx: &CanvasRenderingContext2d, ch_code: u32, x: u32, y: u32);
     // #[wasm_bindgen]
-    pub fn measureText(ctx: &CanvasRenderingContext2d, ch: u32, font_size: u32, name: u32) -> f32;
+    pub fn measureText(ctx: &CanvasRenderingContext2d, ch: u32, font_size: u32, name: f64) -> f32;
     // #[wasm_bindgen]
 	#[wasm_bindgen(catch)]
-	pub async fn loadFile(image_name: u32) -> Result<JsValue, JsValue>;
+	pub async fn loadFile(image_name: f64) -> Result<JsValue, JsValue>;
 	// 加载图片作文canvas
 	#[wasm_bindgen(catch)]
-	pub async fn loadImageAsCanvas(image_name: u32) -> Result<JsValue, JsValue>;
+	pub async fn loadImageAsCanvas(image_name: f64) -> Result<JsValue, JsValue>;
     // 加载ktx文件
     #[wasm_bindgen(catch)]
-	pub async fn loadKtx(image_name: u32) -> Result<JsValue, JsValue>;
+	pub async fn loadKtx(image_name: f64) -> Result<JsValue, JsValue>;
     // 加载图片作文canvas
 	#[wasm_bindgen(catch)]
-	pub async fn loadImage(image_name: u32) -> Result<JsValue, JsValue>;
+	pub async fn loadImage(image_name: f64) -> Result<JsValue, JsValue>;
     // #[wasm_bindgen]
     pub fn useVao() -> bool;
-    pub fn hasAtom(key: u32) -> bool;
-    pub fn setAtom(key: u32, v: String);
-    pub fn int(key: u32, v: String);
+    pub fn hasAtom(key: f64) -> bool;
+    pub fn setAtom(key: f64, v: String);
+    pub fn int(key: f64, v: String);
 
     pub async fn initLocalStore();
     /**
